Login and register components created
