+++ frontend/src/App.js
@@ -1,5 +1,5 @@
 import Menu from './pages/Menu';
-import Login from './pages/Login';
+import Login from './components/Login';
 import Register from './components/Register';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
@@ -9,6 +9,7 @@ export default function App() {
       <Router>
         <Routes>
           <Route path='/register' element={<Register />} />
+          <Route path='/login' element={<Login />} />
           <Route path='/' element={<Login />} />
           <Route path='/menu' element={<Menu />} />
         </Routes>
